==> scripts/05_run_all.R <==
"								
Run All
"

# Set dir 
setwd(githubdir)
setwd("partisan_gap/")

source("scripts/01_j_and_j.R")
source("scripts/02_anes.R")
source("scripts/03_prior_and_bullock.R")
source("scripts/04_density_plot_and_table.R")
